Fix circular dependency and Copilot
